==> 03_EdgeR_DGE.Rmd <==
---
jupyter:
  jupytext:
    formats: ipynb,Rmd
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.1'
      jupytext_version: 1.2.4
  kernelspec:
    display_name: R [conda env:GO]
    language: R
    name: conda-env-GO-r
---

# Honours Project - Differential gene expression (edgeR)
Schurch et al. have shown that for a low number of replicates, edgeR exact is the preferred tool to use.\
Additionally, a fold change threshold of 1 should be applied. (1)
1. Schurch, N. J., Schofield, P., Gierliński, M., Cole, C., Sherstnev, A., Singh, V., … Barton, G. J. (2016).\
How many biological replicates are needed in an RNA-seq experiment and which differential expression tool should you use? RNA, 22(6), 839. https://doi.org/10.1261/RNA.053959.115

```{r}
countdf <- readRDS("data/devspine_counts_excluded.rds")
indTPM <- readRDS("data/TPM.rds")
indTPM = indTPM[,-c(4, 5)]
```

```{r}
#Calculate TPM mean/max:
countdf["TPM12.5"] <- apply(indTPM[,1:3], 1, mean)
countdf["TPMAdult"] <- apply(indTPM[,4:6], 1, mean)
countdf["maxTPM"] <- apply(countdf[,c("TPM12.5","TPMAdult")], 1, max)

#Conduct edgeR analysis:
library(edgeR)
library(stringr)
library(ggplot2)
cond <- str_sub(colnames(countdf[,1:6]),1,5)
dge <- DGEList(countdf[,1:6], group=cond) #create DGEList object containing data from E12.5 and Adult
keep <- filterByExpr(dge) #determine genes with low counts to be excluded
dge <- dge[keep, , keep.lib.sizes=FALSE] #remove genes with low counts
dge <- calcNormFactors(dge, method="TMM") #normalise samples against each other
design <- model.matrix( ~dge$sample$group) #create model (which samples are compared against which)
dge <- estimateGLMCommonDisp(dge, design)
dge <- estimateGLMTrendedDisp(dge, design)
dge <- estimateGLMTagwiseDisp(dge, design)
ext <- exactTest(dge) #apply test
fit = glmFit(dge, design)
all <- topTags(ext, n=nrow(ext$table), adjust.method="BY", sort.by="none")
all = merge(all$table, countdf[,7:9], by=0, all.x=TRUE, all.y=FALSE,sort=FALSE)
names(all)[names(all) == 'y'] <- 'TPM'
sig <- decideTestsDGE(ext, adjust.method="BY", p.value=0.01, lfc=0.5)
sig <- all[sig != 0,]
all <- all[order(all$PValue),]

#plot volcano plot:
volcano <- ggplot(data=all, aes(-logFC, -log2(PValue))) + geom_point(aes(color=log2(maxTPM)), size=1) +
        scale_colour_viridis_c() + #geom_text_repel(data=head(all, 7), aes(label=Row.names)) 
        coord_cartesian(xlim = c(-20, 20)) + geom_hline(yintercept=6.644) + geom_vline(xintercept=0.5) + geom_vline(xintercept=-0.5) +
        coord_cartesian(ylim = c(-10, 1100))
        
volcano
```

```{r}
ggsave(file = "plots/volcano.png", plot = volcano, width = 8, height = 8)
```

A number of genes seem to have extremely low p-values.\
To further investigate these genes, a threshold of -log(PValue) > 1000 was set to catch them.

```{r}
all[-log(all$PValue) > 1000,]
```

A list of 15 genes was identified with a p-value of 0. This makes plotting them awkward. \
Considering how low the next highest p-value is, though, they do not form a unique cluster and shouldn't be treated separately. \
Variation in TPM is quite high in the sample, with apparently few genes reaching the highest values.

```{r}
geneann = readRDS("data/devspine_genes.rds")
annall = merge(all, geneann[,c("ensembl_id","gene_name", "gene_biotype")], by.x="Row.names", by.y="ensembl_id", all.x=TRUE, all.y=FALSE, sort=FALSE)
annall[order(-annall$maxTPM),][1:20,]
```

The genes with very high expression levels appear to be mainly non-protein-coding genes.

```{r}
allproteins = annall[annall$gene_biotype == "protein_coding",]
nrow(allproteins)
nrow(annall)
```

For the sake of further analysis, they were excluded. This reduces the list from 15,476 genes to 13,588 genes.

```{r}
volcano <- ggplot(data=allproteins, aes(-logFC, -log2(PValue))) + geom_point(aes(color=log2(maxTPM)), size=1) +
        scale_colour_viridis_c() + #geom_text_repel(data=head(all, 7), aes(label=Row.names)) 
        coord_cartesian(xlim = c(-20, 20)) + geom_hline(yintercept=6.644) + geom_vline(xintercept=0.5) + geom_vline(xintercept=-0.5) +
        coord_cartesian(ylim = c(-10, 1100))

volcano
```

```{r}
ggsave(file = "plots/volcano_proteins.png", plot = volcano, width = 6, height = 6)
```

Plotting the volcano plot again reveals no big difference, with the exception of expression levels being more clearly distinguished due to the absence of a few extremely highly expressed non-protein-coding genes.

```{r}
sigprot = merge(sig, allproteins, all.x=FALSE, all.y=FALSE, sort=FALSE)
saveRDS(sigprot, "data/sig_regulated_genes.rds")
sum(sigprot$logFC < 0)
sum(sigprot$logFC > 0)
mean(sigprot[sigprot$logFC < 0,]$maxTPM)
mean(sigprot[sigprot$logFC > 0,]$maxTPM)
```

Overall, over 1000 more genes are upregulated than are downregulated in adult cells.\
Additionally, embryonal expression levels of downregulated genes are far lower than adult expression levels of upregulated genes.
